Working mining system and cargo

+++ UnityProject/Assets/Scripts/Runtime/Base.cs
@@ -21,8 +21,8 @@ namespace AC
             _sprite.color = _type.resourceColor;
         }
 
-        public void TryLoadMineral(ResourceDef resourceDef, float amount) => TryLoadMineral(resourceDef.resourceIndex, amount);
-        public void TryLoadMineral(ResourceIndex resourceIndex, float amount)
+        public void TryLoadMineral(ResourceDef resourceDef, int amount) => TryLoadMineral(resourceDef.resourceIndex, amount);
+        public void TryLoadMineral(ResourceIndex resourceIndex, int amount)
         {
             _resources.LoadResource(resourceIndex, amount);
         }
